serializer files were changed

+++ api/views.py
@@ -1,43 +1,81 @@
 from rest_framework.response import Response
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
 
-from userapp.models import CustomUser
+from .serializers import (
+    ServiceSerializer, CustomUserSerializer)
 from serviceapp.models import Service
-
-from .serializers import (CustomUserSerializer, ServiceSerializer)
 
 
 class RoutesView(APIView):
     def get(self, request):
         routes = [
             {"GET": "api/"},
-            {"GET": "api/users/"},
-            {"GET": "api/services/"}
+            {"GET": "api/services/"},
+            {"POST": "api/auth/login/"},
+            {"POST": "api/auth/register/"},
         ]
         return Response(routes)
 
 
-class UserListView(APIView):
+class GetUserView(APIView):
     def get(self, request):
-        users = CustomUser.objects.all()
-        serializer = CustomUserSerializer(users, many=True)
+        user = request.user
+        serializer = CustomUserSerializer(user, many=True)
         return Response(serializer.data)
 
 
-class ServiceListView(APIView):
+class ServiceCRUDView(APIView):
+    permission_classes = [IsAuthenticated, ]
+
     def get(self, request):
+        user = request.user
+        user_services = []
         services = Service.objects.all()
+        # for service in services:
+        #     if user.id == service.user.id:
+        #         user_services.append(service)
+        # if user_services is not None:
+        #     serializer = ServiceSerializer(user_services, many=True)
+        #     return Response(serializer.data)
         serializer = ServiceSerializer(services, many=True)
         return Response(serializer.data)
 
     def post(self, request):
-        pass
+        user = request.user
+        data = request.data
+        if user.is_authenticated:
+            data['user'] = user.id
+            serializer = ServiceSerializer(data=data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response({"Serializer did not pass valid"})
+        return Response({"data": "access denied"})
 
     def patch(self, request):
-        pass
+        user = request.user
+        if user.is_authenticated:
+            data = request.data
+            service = Service.objects.get(id=data['id'])
+            if service.user == user:
+                serializer = ServiceSerializer(service, data=data)
+                if serializer.is_valid():
+                    serializer.save()
+                    return Response(serializer.data)
+                return Response(serializer.errors)
+            return Response({"data": "service is not your"})
+        return Response({"data": "access denied"})
 
     def delete(self, request):
-        deleteId = self.request.query_params("deleteId")
-        service = Service.objects.get(id=deleteId)
-        service.delete()
-
+        user = request.user
+        if user.is_authenticated:
+            deleteId = self.request.query_params.get("deleteId")
+            service = Service.objects.get(id=deleteId)
+            if service is not None:
+                if service.user == user:
+                    service.delete()
+                    return Response({"data": "success"})
+                return Response({"data": "user does not equal"})
+            return Response({"data": "service is unavailable"})
+        return Response({"data": "access denied"})
